Rename classes accroding the filename
